Refactor GUI components into separate modules
